+++ components/confettiCanvasBtn/confetti.helpers.ts
